+++ bin/check-kube-service-available.rb
@@ -16,8 +16,18 @@
 #   gem: kube-client
 #
 # USAGE:
-# -s SERVER - The kube server to use
-# -l SERVICES - The comma delimited list of services to check
+# -s, --api-server URL             URL to API server
+# -v, --api-version VERSION        API version. Defaults to 'v1'
+#     --in-cluster                 Use service account authentication
+#     --ca-file CA-FILE            CA file to verify API server cert
+#     --cert CERT-FILE             Client cert to present
+#     --key KEY-FILE               Client key for the client cert
+# -u, --user USER                  User with access to API
+#     --password PASSWORD          If user is passed, also pass a password
+#     --token TOKEN                Bearer token for authorization
+#     --token-file TOKEN-FILE      File containing bearer token for authorization
+# -l, --list SERVICES              List of services to check (required)
+# -p, --pending SECONDS            Time (in seconds) a pod may be pending for and be valid
 #
 # NOTES:
 #
@@ -27,13 +37,10 @@
 #   for details.
 #
 
-require 'sensu-plugins-kubernetes'
-require 'json'
-require 'kubeclient'
+require 'sensu-plugins-kubernetes/cli'
 require 'time'
 
-class AllServicesUp < Sensu::Plugin::Check::CLI
-
+class AllServicesUp < Sensu::Plugins::Kubernetes::CLI
   @options = Sensu::Plugins::Kubernetes::CLI.options.dup
 
   option :service_list,
@@ -50,10 +57,7 @@ class AllServicesUp < Sensu::Plugin::Check::CLI
          proc: proc(&:to_i)
 
   def run
-    cli = AllServicesUp.new
-    client = self.get_client(cli)
-
-    services = parse_list(cli.config[:service_list])
+    services = parse_list(config[:service_list])
     failed_services = []
     s = client.get_services
     s.each do |a|
@@ -78,7 +82,7 @@ class AllServicesUp < Sensu::Plugin::Check::CLI
         case p.status.phase
         when 'Pending'
           next if p.status.startTime.nil?
-          if (Time.now - Time.parse(p.status.startTime)).to_i < cli.config[:pendingTime]
+          if (Time.now - Time.parse(p.status.startTime)).to_i < config[:pendingTime]
             pod_available = True
             break
           end
@@ -105,6 +109,8 @@ class AllServicesUp < Sensu::Plugin::Check::CLI
     else
       critical "Some services could not be checked: #{services.join(' ')}"
     end
+  rescue KubeException => e
+    critical 'API error: ' << e.message
   end
 
   def parse_list(list)